bot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
discord.py==2.4.0
pydantic-settings==2.6.1
httpx==0.27.2
